(PC-17175) feat(cookies): remove FF appEnableCookiesV2
