+++ src/features/cookies/helpers/useCookies.test.ts
@@ -27,11 +27,6 @@ const mockStartTrackingAcceptedCookies = jest.spyOn(
   'startTrackingAcceptedCookies'
 )
 
-const mockSettings = jest.fn().mockReturnValue({ data: { appEnableCookiesV2: true } })
-jest.mock('features/auth/settings', () => ({
-  useAppSettings: jest.fn(() => mockSettings()),
-}))
-
 const COOKIES_CONSENT_KEY = 'cookies'
 const deviceId = 'testUuidV4'
 const TODAY = new Date('2022-09-29')
@@ -489,72 +484,6 @@ describe('useCookies', () => {
         accepted: [],
         refused: ALL_OPTIONAL_COOKIES,
       },
-    })
-  })
-
-  describe('Feature Flag', () => {
-    describe.each([
-      ['undefined', {}],
-      ['disabled', { appEnableCookiesV2: false }],
-    ])('when feature flag is %s', (_, settings) => {
-      beforeEach(() => mockSettings.mockReturnValueOnce({ data: settings }))
-
-      it('should not change cookies consent', async () => {
-        storage.saveObject(COOKIES_CONSENT_KEY, {
-          buildVersion: Package.build,
-          deviceId,
-          choiceDatetime: TODAY.toISOString(),
-          consent: {
-            mandatory: COOKIES_BY_CATEGORY.essential,
-            accepted: ALL_OPTIONAL_COOKIES,
-            refused: [],
-          },
-        })
-
-        renderHook(useCookies)
-        await superFlushWithAct()
-
-        expect(mockStartTrackingAcceptedCookies).not.toHaveBeenCalled()
-      })
-
-      it('should not save cookies consent in the storage', async () => {
-        const { result } = renderHook(useCookies)
-        const { setCookiesConsent } = result.current
-
-        await act(async () => {
-          await setCookiesConsent({
-            mandatory: COOKIES_BY_CATEGORY.essential,
-            accepted: [],
-            refused: ALL_OPTIONAL_COOKIES,
-          })
-        })
-
-        const cookiesConsent = await storage.readObject(COOKIES_CONSENT_KEY)
-        expect(cookiesConsent).toBeNull()
-      })
-
-      it('should not set user ID', async () => {
-        const cookiesConsentWithoutUserId = {
-          buildVersion: Package.build,
-          deviceId,
-          choiceDatetime: TODAY.toISOString(),
-          consent: {
-            mandatory: COOKIES_BY_CATEGORY.essential,
-            accepted: ALL_OPTIONAL_COOKIES,
-            refused: [],
-          },
-        }
-        storage.saveObject(COOKIES_CONSENT_KEY, cookiesConsentWithoutUserId)
-        const { result } = renderHook(useCookies)
-        const { setUserId } = result.current
-
-        await act(async () => {
-          await setUserId(FAKE_USER_ID)
-        })
-
-        const cookiesConsent = await storage.readObject(COOKIES_CONSENT_KEY)
-        expect(cookiesConsent).toEqual(cookiesConsentWithoutUserId)
-      })
     })
   })
 })
