+++ src/features/cookies/pages/CookiesDetails.native.test.tsx
@@ -3,11 +3,6 @@ import React from 'react'
 import { CookiesDetails } from 'features/cookies/pages/CookiesDetails'
 import { flushAllPromisesWithAct, render } from 'tests/utils'
 
-const mockSettings = jest.fn().mockReturnValue({ data: { appEnableCookiesV2: true } })
-jest.mock('features/auth/settings', () => ({
-  useAppSettings: jest.fn(() => mockSettings()),
-}))
-
 jest.mock('features/profile/api')
 
 describe('<CookiesDetails/>', () => {
