+++ src/features/firstLogin/PrivacyPolicy/PrivacyPolicy.tsx
@@ -1,13 +1,42 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 
-import { useAppSettings } from 'features/auth/settings'
-import { PrivacyPolicyV1 } from 'features/firstLogin/PrivacyPolicy/PrivacyPolicyV1'
-import { PrivacyPolicyV2 } from 'features/firstLogin/PrivacyPolicy/PrivacyPolicyV2'
+import { ConsentState } from 'features/cookies/enums'
+import { useCookies } from 'features/cookies/helpers/useCookies'
+import { useIsCookiesListUpToDate } from 'features/cookies/helpers/useIsCookiesListUpToDate'
+import { CookiesConsent } from 'features/cookies/pages/CookiesConsent'
+import { useModal } from 'ui/components/modals/useModal'
 
 export function PrivacyPolicy() {
-  const { data: settings } = useAppSettings()
+  const { cookiesConsent: hasUserMadeCookieChoiceV2 } = useCookies()
+  const {
+    visible: cookiesConsentModalVisible,
+    hideModal: hideCookiesConsentModal,
+    showModal: showCookiesConsentModal,
+  } = useModal(false)
+  const isCookiesListUpToDate = useIsCookiesListUpToDate()
 
-  if (settings === undefined) return null
+  useEffect(() => {
+    switch (hasUserMadeCookieChoiceV2.state) {
+      case ConsentState.LOADING:
+        break
 
-  return settings.appEnableCookiesV2 ? <PrivacyPolicyV2 /> : <PrivacyPolicyV1 />
+      case ConsentState.UNKNOWN:
+        showCookiesConsentModal()
+        break
+
+      default:
+        if (isCookiesListUpToDate) {
+          hideCookiesConsentModal()
+        } else {
+          showCookiesConsentModal()
+        }
+    }
+  }, [
+    hasUserMadeCookieChoiceV2,
+    hideCookiesConsentModal,
+    isCookiesListUpToDate,
+    showCookiesConsentModal,
+  ])
+
+  return <CookiesConsent visible={cookiesConsentModalVisible} hideModal={hideCookiesConsentModal} />
 }
