+++ src/features/navigation/__tests__/RootNavigator.web.test.tsx
@@ -5,7 +5,6 @@ import { mocked } from 'ts-jest/utils'
 import { useMustUpdateApp } from 'features/forceUpdate/useMustUpdateApp'
 import { useCurrentRoute } from 'features/navigation/helpers'
 import { useSplashScreenContext } from 'libs/splashscreen'
-import { storage } from 'libs/storage'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { flushAllPromisesWithAct, render } from 'tests/utils/web'
 
@@ -14,9 +13,6 @@ import { RootNavigator } from '../RootNavigator'
 const mockUseSplashScreenContext = mocked(useSplashScreenContext)
 const mockedUseMustUpdateApp = mocked(useMustUpdateApp)
 const mockUseCurrentRoute = mocked(useCurrentRoute)
-const mockSettings = jest.fn().mockReturnValue({
-  data: { appEnableCookiesV2: false },
-})
 
 jest.mock('features/cookies/helpers/useIsCookiesListUpToDate')
 jest.mock('features/forceUpdate/useMustUpdateApp')
@@ -35,18 +31,11 @@ jest.mock('features/navigation/RootNavigator/useInitialScreenConfig', () => ({
   useInitialScreen: () => 'TabNavigator',
 }))
 jest.mock('features/navigation/helpers')
-jest.mock('features/auth/settings', () => ({
-  useAppSettings: jest.fn(() => mockSettings()),
-}))
 jest.mock('libs/splashscreen')
 
 describe('<RootNavigator />', () => {
   beforeEach(() => {
     mockUseCurrentRoute.mockReturnValue({ name: 'TabNavigator', key: 'key' })
-  })
-
-  afterEach(async () => {
-    await storage.clear('has_accepted_cookie')
   })
 
   it('should NOT display PrivacyPolicy if splash screen is not yet hidden', async () => {
@@ -90,13 +79,11 @@ describe('<RootNavigator />', () => {
 
 describe('ForceUpdate display logic', () => {
   it('should display force update page when global variable is set', async () => {
-    await storage.saveObject('has_accepted_cookie', false)
     mockUseCurrentRoute.mockReturnValueOnce({ name: 'TabNavigator', key: 'key' })
     mockedUseMustUpdateApp.mockReturnValueOnce(true)
 
     const rootNavigator = await renderRootNavigator()
 
-    expect(rootNavigator).toMatchSnapshot()
     expect(rootNavigator.queryAllByText("Mise à jour de l'application")).toBeTruthy()
   })
 })
