+++ src/features/profile/pages/ConsentSettings/ConsentSettings.native.test.tsx
@@ -1,184 +1,143 @@
-import { Route } from '@react-navigation/native'
-import { StackScreenProps } from '@react-navigation/stack'
+import mockdate from 'mockdate'
 import React from 'react'
 
-import { goBack } from '__mocks__/@react-navigation/native'
-import { openUrl } from 'features/navigation/helpers/openUrl'
-import { usePreviousRoute } from 'features/navigation/helpers/usePreviousRoute'
-import { RootStackParamList } from 'features/navigation/RootNavigator'
-import { env } from 'libs/environment'
+import Package from '__mocks__/package.json'
+import { COOKIES_BY_CATEGORY } from 'features/cookies/CookiesPolicy'
+import * as Tracking from 'features/cookies/helpers/startTrackingAcceptedCookies'
+import { ConsentSettings } from 'features/profile/pages/ConsentSettings/ConsentSettings'
 import { analytics } from 'libs/firebase/analytics'
 import { storage } from 'libs/storage'
-import { superFlushWithAct, render, fireEvent, waitFor } from 'tests/utils'
+import { requestIDFATrackingConsent } from 'libs/trackingConsent/useTrackingConsent'
+import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
+import { fireEvent, render, waitFor } from 'tests/utils'
+import { SNACK_BAR_TIME_OUT } from 'ui/components/snackBar/SnackBarContext'
+import { SnackBarHelperSettings } from 'ui/components/snackBar/types'
 
-import { ConsentSettings } from './ConsentSettings'
+const COOKIES_CONSENT_KEY = 'cookies'
+const Today = new Date(2022, 9, 29)
+mockdate.set(Today)
+const deviceId = 'testUuidV4'
 
-jest.mock('features/auth/settings')
-jest.mock('features/navigation/helpers/openUrl')
-const mockedOpenUrl = openUrl as jest.MockedFunction<typeof openUrl>
+const mockNavigate = jest.fn()
+jest.mock('@react-navigation/native', () => ({
+  ...jest.requireActual('@react-navigation/native'),
+  useNavigation: () => ({ navigate: mockNavigate, push: jest.fn() }),
+  useFocusEffect: jest.fn(),
+}))
 
-jest.mock('features/navigation/helpers/usePreviousRoute')
-const mockedUsePreviousRoute = usePreviousRoute as jest.MockedFunction<typeof usePreviousRoute>
+jest.mock('libs/trackingConsent/useTrackingConsent')
+const mockrequestIDFATrackingConsent = requestIDFATrackingConsent as jest.Mock
+
+const mockStartTrackingAcceptedCookies = jest.spyOn(Tracking, 'startTrackingAcceptedCookies')
 
 const mockShowSuccessSnackBar = jest.fn()
 jest.mock('ui/components/snackBar/SnackBarContext', () => ({
   useSnackBarContext: () => ({
-    showSuccessSnackBar: mockShowSuccessSnackBar,
+    showSuccessSnackBar: jest.fn((props: SnackBarHelperSettings) => mockShowSuccessSnackBar(props)),
   }),
 }))
 
-// eslint-disable-next-line local-rules/no-allow-console
-allowConsole({ error: true })
-
-describe('ConsentSettings', () => {
-  beforeEach(() => {
-    storage.clear('has_accepted_cookie')
+describe('<ConsentSettings/>', () => {
+  it('should render correctly', () => {
+    const renderAPI = renderConsentSettings()
+    expect(renderAPI).toMatchSnapshot()
   })
 
-  it('should display inactive switch et set has_accepted_cookie to false if has_accepted_cookie is null', async () => {
-    storage.saveObject('has_accepted_cookie', null)
-    const { expectTrackingSwitchDisabled } = await renderConsentSettings()
+  it('should persist cookies consent information when user partially accepts cookies', async () => {
+    const { getByText, getByTestId } = renderConsentSettings()
 
+    const performanceSwitch = getByTestId('Interrupteur-performance')
+    fireEvent.press(performanceSwitch)
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
+
+    const storageContent = {
+      buildVersion: Package.build,
+      deviceId,
+      choiceDatetime: Today.toISOString(),
+      consent: {
+        mandatory: COOKIES_BY_CATEGORY.essential,
+        accepted: COOKIES_BY_CATEGORY.performance,
+        refused: [...COOKIES_BY_CATEGORY.customization, ...COOKIES_BY_CATEGORY.marketing],
+      },
+    }
     await waitFor(async () => {
-      expect(await storage.readObject('has_accepted_cookie')).toBe(false)
-      expectTrackingSwitchDisabled()
+      expect(await storage.readObject(COOKIES_CONSENT_KEY)).toEqual(storageContent)
     })
   })
 
-  it('should display inactive switch if user choise is false', async () => {
-    storage.saveObject('has_accepted_cookie', false)
-    const { expectTrackingSwitchDisabled } = await renderConsentSettings()
+  it('should request tracking transparency when user saves cookies choice', async () => {
+    const { getByText } = renderConsentSettings()
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
 
     await waitFor(() => {
-      expectTrackingSwitchDisabled()
+      expect(mockrequestIDFATrackingConsent).toHaveBeenCalled()
     })
   })
 
-  it('should display active switch if user choise is true', async () => {
-    storage.saveObject('has_accepted_cookie', true)
-    const { expectTrackingSwitchEnabled } = await renderConsentSettings()
+  it('should call startTrackingAcceptedCookies with empty array if user refuses all cookies', async () => {
+    const { getByText } = renderConsentSettings()
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
 
     await waitFor(() => {
-      expectTrackingSwitchEnabled()
+      expect(mockStartTrackingAcceptedCookies).toHaveBeenCalledWith([])
     })
   })
 
-  it('should enable save button on toggle switch', async () => {
-    const { pressTrackingSwitch, expectSaveButtonEnabled } = await renderConsentSettings()
+  it('should call startTrackingAcceptedCookies with cookies performance if user accepts performance cookies', async () => {
+    const { getByTestId, getByText } = renderConsentSettings()
 
-    pressTrackingSwitch()
+    const performanceSwitch = getByTestId('Interrupteur-performance')
+    fireEvent.press(performanceSwitch)
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
 
     await waitFor(() => {
-      expectSaveButtonEnabled()
+      expect(mockStartTrackingAcceptedCookies).toHaveBeenCalledWith(COOKIES_BY_CATEGORY.performance)
     })
   })
 
-  it('should save user choice in storage and call setAnalyticsCollectionEnabled on press save', async () => {
-    const {
-      expectTrackingSwitchEnabled,
-      expectSaveButtonEnabled,
-      pressTrackingSwitch,
-      pressSaveButton,
-    } = await renderConsentSettings()
+  it('should log analytics if performance cookies are accepted', async () => {
+    const { getByText, getByTestId } = renderConsentSettings()
 
-    pressTrackingSwitch()
+    const performanceSwitch = getByTestId('Interrupteur-performance')
+    fireEvent.press(performanceSwitch)
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
+
+    await waitFor(() =>
+      expect(analytics.logHasMadeAChoiceForCookies).toHaveBeenCalledWith({
+        from: 'ConsentSettings',
+        type: { performance: true, customization: false, marketing: false },
+      })
+    )
+  })
+
+  it('should show snackbar and navigate to home on save', async () => {
+    const { getByText } = renderConsentSettings()
+
+    const saveChoice = getByText('Enregistrer mes choix')
+    fireEvent.press(saveChoice)
 
     await waitFor(() => {
-      expectTrackingSwitchEnabled()
-      expectSaveButtonEnabled()
+      expect(mockShowSuccessSnackBar).toBeCalledWith({
+        message: 'Ton choix a bien été enregistré.',
+        timeout: SNACK_BAR_TIME_OUT,
+      })
+      expect(mockNavigate).toBeCalledWith('TabNavigator', { screen: 'Profile' })
     })
-
-    pressSaveButton()
-
-    await waitFor(async () => {
-      expect(await storage.readObject('has_accepted_cookie')).toBe(true)
-      expect(analytics.enableCollection).toHaveBeenCalled()
-    })
-  })
-
-  it('should go back on press save', async () => {
-    const { getByTestId, getByText } = await renderConsentSettings()
-
-    const toggleButton = getByTestId('Interrupteur')
-    fireEvent.press(toggleButton)
-
-    await waitFor(() => {
-      const saveButton = getByText('Enregistrer')
-      fireEvent.press(saveButton)
-      expect(goBack).toHaveBeenCalled()
-    })
-  })
-
-  it('should show snackbar on press save', async () => {
-    const { getByTestId, getByText } = await renderConsentSettings()
-
-    const toggleButton = getByTestId('Interrupteur')
-    fireEvent.press(toggleButton)
-
-    await waitFor(() => {
-      const saveButton = getByText('Enregistrer')
-      fireEvent.press(saveButton)
-      expect(mockShowSuccessSnackBar).toHaveBeenCalled()
-    })
-  })
-
-  it('should open cookies policies on click on "Politique des cookies"', async () => {
-    const { getByText } = await renderConsentSettings()
-
-    fireEvent.press(getByText('Politique des cookies'))
-
-    await superFlushWithAct(1)
-    expect(mockedOpenUrl).toBeCalledWith(env.COOKIES_POLICY_LINK, undefined)
   })
 })
 
-const TRACKING_SWITCH_TEST_ID = 'Interrupteur'
-const SAVE_BUTTON_TEST_ID = 'Enregistrer'
-
-async function renderConsentSettings() {
-  mockedUsePreviousRoute.mockReturnValue(jest.fn() as unknown as Route<string>)
-
-  const onGoBack = jest.fn() as () => void
-  const navigationProps = {
-    route: {
-      params: { onGoBack },
-    },
-  } as StackScreenProps<RootStackParamList, 'ConsentSettings'>
-
-  const renderAPI = render(<ConsentSettings {...navigationProps} />)
-  await superFlushWithAct(1)
-
-  function pressTrackingSwitch() {
-    const trackingSwitch = renderAPI.getByTestId(TRACKING_SWITCH_TEST_ID)
-    fireEvent.press(trackingSwitch)
-  }
-
-  function expectTrackingSwitchEnabled() {
-    const trackingSwitch = renderAPI.getByTestId(TRACKING_SWITCH_TEST_ID)
-    expect(trackingSwitch.parent?.props.accessibilityState.checked).toBeTruthy()
-  }
-
-  function expectTrackingSwitchDisabled() {
-    const trackingSwitch = renderAPI.getByTestId(TRACKING_SWITCH_TEST_ID)
-    expect(trackingSwitch.parent?.props.accessibilityState.checked).toBeFalsy()
-  }
-
-  function pressSaveButton() {
-    const saveButton = renderAPI.getByTestId(SAVE_BUTTON_TEST_ID)
-    fireEvent.press(saveButton)
-  }
-
-  function expectSaveButtonEnabled() {
-    const saveButton = renderAPI.getByTestId(SAVE_BUTTON_TEST_ID)
-    expect(saveButton).toBeEnabled()
-  }
-
-  return {
-    ...renderAPI,
-    expectTrackingSwitchEnabled,
-    expectTrackingSwitchDisabled,
-    expectSaveButtonEnabled,
-    pressTrackingSwitch,
-    pressSaveButton,
-  }
-}
+const renderConsentSettings = () =>
+  render(<ConsentSettings />, {
+    // eslint-disable-next-line local-rules/no-react-query-provider-hoc
+    wrapper: ({ children }) => reactQueryProviderHOC(children),
+  })
