+++ src/libs/algolia/analytics/logClickOnOffer.ts
@@ -1,21 +1,18 @@
 import { useMemo } from 'react'
 import AlgoliaSearchInsights from 'search-insights'
 
-import { useAppSettings } from 'features/auth/settings'
 import { CookieNameEnum } from 'features/cookies/enums'
 import { getAcceptedCookieConsent } from 'features/cookies/helpers/getAcceptedCookieConsent'
 import { useSearchAnalyticsState } from 'libs/algolia/analytics/SearchAnalyticsWrapper'
 import { env } from 'libs/environment'
 import { captureMonitoringError } from 'libs/monitoring'
-import { getCookiesConsent } from 'libs/trackingConsent/consent'
 
 export const logClickOnOffer =
-  (appEnableCookiesV2: boolean, currentQueryID?: string) =>
+  (currentQueryID?: string) =>
   async ({ objectID, position }: { objectID: string; position: number }) => {
-    // TODO(PC-17175): use getAcceptedCookieConsent instead
-    const hasAcceptedAlgoliaInsights = appEnableCookiesV2
-      ? await getAcceptedCookieConsent(CookieNameEnum.ALGOLIA_INSIGHTS)
-      : await getCookiesConsent()
+    const hasAcceptedAlgoliaInsights = await getAcceptedCookieConsent(
+      CookieNameEnum.ALGOLIA_INSIGHTS
+    )
 
     if (!hasAcceptedAlgoliaInsights) return
 
@@ -36,13 +33,5 @@ export const logClickOnOffer =
 
 export const useLogClickOnOffer = () => {
   const { currentQueryID } = useSearchAnalyticsState()
-  const { data: settings } = useAppSettings()
-  const appEnableCookiesV2 = !!settings?.appEnableCookiesV2
-
-  return useMemo(
-    () => ({
-      logClickOnOffer: logClickOnOffer(appEnableCookiesV2, currentQueryID),
-    }),
-    [appEnableCookiesV2, currentQueryID]
-  )
+  return useMemo(() => ({ logClickOnOffer: logClickOnOffer(currentQueryID) }), [currentQueryID])
 }
