--- src/libs/trackingConsent/index.ts
@@ -1,1 +1,0 @@
-export { useTrackingConsent } from './useTrackingConsent'
