+++ src/libs/utm/getUtmParamsConsent.test.ts
@@ -6,11 +6,6 @@ import { storage } from 'libs/storage'
 import { getUtmParamsConsent } from 'libs/utm/getUtmParamsConsent'
 import { act, flushAllPromisesWithAct, renderHook } from 'tests/utils'
 
-const mockSettings = jest.fn().mockReturnValue({ data: { appEnableCookiesV2: true } })
-jest.mock('features/auth/settings', () => ({
-  useAppSettings: jest.fn(() => mockSettings()),
-}))
-
 jest.mock('features/profile/api')
 jest.mock('api/api')
