+++ src/libs/utm/storeUtmParams.test.ts
@@ -7,11 +7,6 @@ import { storage } from 'libs/storage'
 import { storeUtmParams } from 'libs/utm/storeUtmParams'
 import { act, flushAllPromisesWithAct, renderHook } from 'tests/utils'
 
-const mockSettings = jest.fn().mockReturnValue({ data: { appEnableCookiesV2: true } })
-jest.mock('features/auth/settings', () => ({
-  useAppSettings: jest.fn(() => mockSettings()),
-}))
-
 jest.mock('features/profile/api')
 jest.mock('api/api')
